backend/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.6
pymongo==3.12.0 